parametrized slow tests
