+++ tardis/io/tests/data/tardis_w7_13d_densities.dat
@@ -1,0 +1,31 @@
+   0  11000.000 7.1222555e-14
+   1  11300.000 7.1222555e-14
+   2  11600.000 6.3072282e-14
+   3  11900.000 5.7007079e-14
+   4  12200.000 5.3086730e-14
+   5  12500.000 5.2311961e-14
+   6  12800.000 5.6781610e-14
+   7  13100.000 5.8228905e-14
+   8  13400.000 5.2054802e-14
+   9  13700.000 4.4574440e-14
+  10  14000.000 3.8773531e-14
+  11  14300.000 4.7217992e-14
+  12  14600.000 4.6619456e-14
+  13  14900.000 3.0112290e-14
+  14  15200.000 1.0403662e-14
+  15  15500.000 7.2243094e-15
+  16  15800.000 5.4831153e-15
+  17  16100.000 4.8490603e-15
+  18  16400.000 3.2750774e-15
+  19  16700.000 3.1899090e-15
+  20  17000.000 3.1062706e-15
+  21  17300.000 2.3697649e-15
+  22  17600.000 2.1888425e-15
+  23  17900.000 2.1773350e-15
+  24  18200.000 2.0463255e-15
+  25  18500.000 1.7174255e-15
+  26  18800.000 1.6454979e-15
+  27  19100.000 1.3374691e-15
+  28  19400.000 1.2110850e-15
+  29  19700.000 1.0728003e-15
+  30  20000.000 9.3758061e-16
